requirements.txt
# AutoGen and AI Model Dependencies  
autogen-agentchat>=0.7.2
autogen-core>=0.7.2
autogen-ext[openai]>=0.7.2
pyautogen==0.10.0

# AI Provider SDKs
openai>=1.99.0
groq==0.31.0
google-generativeai>=0.8.5
anthropic>=0.61.0

# Data and visualization
matplotlib==3.9.4
pandas==2.2.4

# Utility packages
python-dotenv>=1.0.0
aiofiles>=24.1.0
tiktoken>=0.8.0

# Optional: For enhanced functionality
requests>=2.32.0
httpx>=0.28.0
pydantic>=2.11.0
